Add collection data on reef endpoint
