+++ packages/api/src/collections/collections.service.ts
@@ -5,32 +5,18 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { camelCase, isUndefined, keyBy, omitBy } from 'lodash';
-import { In, Repository } from 'typeorm';
+import { isUndefined, omitBy } from 'lodash';
+import { Repository } from 'typeorm';
 import { Collection } from './collections.entity';
 import { DailyData } from '../reefs/daily-data.entity';
-import { Sources, SourceType } from '../reefs/sources.entity';
+import { Sources } from '../reefs/sources.entity';
 import { LatestData } from '../time-series/latest-data.entity';
-import { Metric } from '../time-series/metrics.entity';
 import { User } from '../users/users.entity';
-import { getSstAnomaly } from '../utils/liveData';
 import { hasHoboDataSubQuery } from '../utils/reef.utils';
-import { CollectionDataDto } from './dto/collection-data.dto';
 import { CreateCollectionDto } from './dto/create-collection.dto';
 import { FilterCollectionDto } from './dto/filter-collection.dto';
 import { UpdateCollectionDto } from './dto/update-collection.dto';
-
-interface LatestDailyData {
-  // reefId
-  id: number;
-  date: string;
-  // satelliteTemperature
-  sst?: number;
-  // degreeHeatingStress
-  dhd?: number;
-  // weekly alert
-  alert?: number;
-}
+import { getCollectionData } from '../utils/collections.utils';
 
 @Injectable()
 export class CollectionsService {
@@ -119,7 +105,29 @@ export class CollectionsService {
       return collection;
     }
 
-    return this.getCollectionData(collection);
+    const mappedReefData = await getCollectionData(
+      collection.reefs,
+      this.latestDataRepository,
+      this.dailyDataRepository,
+    );
+
+    const hasHoboData = await hasHoboDataSubQuery(this.sourcesRepository);
+
+    return {
+      ...collection,
+      user: {
+        ...collection.user,
+        firebaseUid: undefined,
+      },
+      reefs: collection.reefs.map((reef) => {
+        return {
+          ...reef,
+          hasHobo: hasHoboData.has(reef.id),
+          applied: reef.applied,
+          collectionData: mappedReefData[reef.id],
+        };
+      }),
+    };
   }
 
   async update(collectionId: number, updateCollectionDto: UpdateCollectionDto) {
@@ -170,89 +178,5 @@ export class CollectionsService {
         `Collection with ID ${collectionId} not found.`,
       );
     }
-  }
-
-  private async getCollectionData(collection: Collection): Promise<Collection> {
-    // Get buoy data
-    const latestData = await this.latestDataRepository.find({
-      where: {
-        reef: In(collection.reefs.map((reef) => reef.id)),
-        source: SourceType.SPOTTER,
-        metric: In([Metric.BOTTOM_TEMPERATURE, Metric.TOP_TEMPERATURE]),
-      },
-    });
-
-    const mappedlatestData = latestData.reduce<
-      Record<number, Record<'bottomTemperature' | 'topTemperature', number>>
-    >((acc, data) => {
-      return {
-        ...acc,
-        [data.reefId]: {
-          ...acc[data.reefId],
-          [camelCase(data.metric)]: data.value,
-        },
-      };
-    }, {});
-
-    // Get latest sst and degree_heating days
-    // Query builder doesn't apply correctly the select and DISTINCT must be first
-    // So we use a raw query to achieve this
-    const latestDailyData: LatestDailyData[] = await this.dailyDataRepository
-      .createQueryBuilder('dailyData')
-      .select(
-        'DISTINCT ON (reef_id) reef_id AS id, satellite_temperature sst, degree_heating_days dhd, weekly_alert_level alert, date',
-      )
-      .where('reef_id IN (:...reefIds)', {
-        reefIds: collection.reefs.map((reef) => reef.id),
-      })
-      .orderBy('reef_id, date', 'DESC')
-      .getRawMany();
-
-    const mappedLatestDailyData: Record<number, LatestDailyData> = keyBy(
-      latestDailyData,
-      'id',
-    );
-
-    const mappedReefData = collection.reefs.reduce<
-      Record<number, CollectionDataDto>
-    >((acc, reef) => {
-      const sstValue = mappedLatestDailyData[reef.id]?.sst;
-      const sst =
-        sstValue || sstValue === 0
-          ? {
-              value: sstValue,
-              timestamp: mappedLatestDailyData[reef.id].date,
-            }
-          : undefined;
-
-      return {
-        ...acc,
-        [reef.id]: {
-          ...mappedlatestData[reef.id],
-          satelliteTemperature: mappedLatestDailyData[reef.id]?.sst,
-          degreeHeatingDays: mappedLatestDailyData[reef.id]?.dhd,
-          weeklyAlertLevel: mappedLatestDailyData[reef.id]?.alert,
-          sstAnomaly: getSstAnomaly(reef.historicalMonthlyMean, sst),
-        },
-      };
-    }, {});
-
-    const hasHoboData = await hasHoboDataSubQuery(this.sourcesRepository);
-
-    return {
-      ...collection,
-      user: {
-        ...collection.user,
-        firebaseUid: undefined,
-      },
-      reefs: collection.reefs.map((reef) => {
-        return {
-          ...reef,
-          hasHobo: hasHoboData.has(reef.id),
-          applied: reef.applied,
-          collectionData: mappedReefData[reef.id],
-        };
-      }),
-    };
   }
 }
